updated service worker handling
