+++ lib/ugma_today.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_redux/flutter_redux.dart';
+import 'package:redux/redux.dart';
+import 'package:ugma_today/lang/delegates.dart';
+import 'package:ugma_today/lang/supported_locales.dart';
+import 'package:ugma_today/redux/states/app_state.dart';
+import 'package:ugma_today/routes/routes.dart';
+
+class UgmaToday extends StatelessWidget {
+  final Store<AppState> store;
+  UgmaToday({Key key, this.store}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return StoreProvider<AppState>(
+      store: this.store,
+      child: MaterialApp(
+        title: 'Ugma Today',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+          visualDensity: VisualDensity.adaptivePlatformDensity,
+        ),
+        initialRoute: '/',
+        routes: routes(),
+        localizationsDelegates: delegates,
+        supportedLocales: locales,
+      ),
+    );
+  }
+}
