+++ web/index.html
@@ -78,11 +78,24 @@
        application. For more information, see:
        https://developers.google.com/web/fundamentals/primers/service-workers -->
   <script>
-    if ('serviceWorker' in navigator) {
-      window.addEventListener('load', function () {
-        navigator.serviceWorker.register('flutter_service_worker.js');
-      });
-    }
+     if ('serviceWorker' in navigator) {
+        window.addEventListener('load', function () {
+          navigator.serviceWorker.register('flutter_service_worker.js').then((reg) => {
+            reg.addEventListener("updatefound", () => {
+             const installingWorker = reg.installing;
+              installingWorker.onstatechange = () => {
+                switch (installingWorker.state) {
+                  case 'installed':
+                    if (navigator.serviceWorker.controller) {
+                      installingWorker.postMessage('skipWaiting');
+                    }
+                    break;
+                }
+              }
+            });
+          });
+        });
+      }
   </script>
   <script src="main.dart.js" type="application/javascript"></script>
 </body>
